top-down/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
boto=2.49.0=pypi_0
boto3=1.10.47=pypi_0
botocore=1.13.47=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2019.11.27=0
cairo=1.14.12=h8948797_3
certifi=2019.11.28=py37_0
cffi=1.13.2=py37h2e261b9_0
chardet=3.0.4=pypi_0
click=7.0=pypi_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.6.5=cuda9.0_0
cycler=0.10.0=py37_0
dbus=1.13.12=h746ee38_0
docutils=0.15.2=pypi_0
easydict=1.9=py_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
gensim=3.8.1=pypi_0
glib=2.63.1=h5a9c865_0
graphite2=1.3.13=h23475e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
intel-openmp=2019.4=243
jasper=2.0.14=h07fcdf6_1
jmespath=0.9.4=pypi_0
joblib=0.14.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=py37he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
matplotlib=3.1.1=py37h5429711_0
mkl=2018.0.3=1
mkl_fft=1.0.6=py37h7dd41cf_0
mkl_random=1.0.1=py37h4414c95_1
nccl=1.3.5=cuda9.0_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
nltk=3.3.0=py37_0
numpy=1.15.4=py37h1d66e8a_0
numpy-base=1.15.4=py37h81de0dd_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1d=h7b6447c_3
pcre=8.43=he6710b0_0
pillow=6.2.1=py37h34e0f95_0
pip=19.3.1=py37_0
pixman=0.38.0=h7b6447c_0
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.19=py37_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.5=h0371630_0
python-dateutil=2.8.1=py_0
pytorch=0.4.1=py37ha74772b_0
pytz=2019.3=py_0
pyyaml=5.2=py37h7b6447c_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
s3transfer=0.2.1=pypi_0
scikit-learn=0.22.1=pypi_0
scipy=1.4.1=pypi_0
setuptools=42.0.2=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
smart-open=1.9.0=pypi_0
sqlite=3.30.1=h7b6447c_0
tbb=2019.8=hfd86e86_0
tbb4py=2019.8=py37hfd86e86_0
theano=1.0.4=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.2.1=py37_0
tornado=6.0.3=py37h7b6447c_0
tqdm=4.26.0=py37h28b3542_0
urllib3=1.25.7=pypi_0
wheel=0.33.6=py37_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
